requirements.txt
APScheduler==3.6.3
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
six==1.15.0
speedtest-cli==2.1.2
toml==0.10.2
tzlocal==2.1
urllib3==1.26.2
wrapt==1.12.1
